// src/app/core/api/models/event-detail.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EventDetail {
  coverPicture?: string;
  creatorName?: string;
  eventBudget?: number;
  eventDate?: string;
  eventDescription?: string;
  eventLocation?: string;
  eventTitle?: string;
  eventType?: 'INTERNAL';
  id?: string;
  isDraft?: boolean;
  totalExpenditure?: number;
}